+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## raspberrypi-ua-netinst v2.1.1
 
 - added installer output via telnet
+- added or changed configuration variables
+  - disable telnet output
 
 ## raspberrypi-ua-netinst v2.1.0